完善模糊查询以及重大删除确认功能

+++ dormitory_manage/bedroom_delete.cs
@@ -19,7 +19,7 @@ namespace dormitory_manage
 
         private void select_Click(object sender, EventArgs e)//从数据库中查询寝室
         {
-            string sql = string.Format("select * from 公寓寝室 where 寝室号='{0}'", textBox1.Text);
+            string sql = string.Format("select * from 公寓寝室 where 寝室号 like '%{0}%'", textBox1.Text);
             dataGridView1.DataSource = getData(sql);
         }
         DataTable getData(string sql)
@@ -30,18 +30,22 @@ namespace dormitory_manage
 
         private void delete_Click(object sender, EventArgs e)//将查询到的寝室删除
         {
-            string sql = string.Format("delete  from 公寓寝室 where 公寓号='{0}' and 寝室号='{1}'", dataGridView1.SelectedCells[0].Value.ToString(),dataGridView1.SelectedCells[1].Value.ToString());
-            int i = Sqlhelper.exexute(sql);
-            if (i == 1)
+            if (MessageBox.Show("您真的要删除吗？", "此删除不可恢复", MessageBoxButtons.YesNo) == DialogResult.Yes)
             {
-                MessageBox.Show("删除成功！");
-                string sql2 = string.Format("select * from 公寓寝室 where 寝室号='{0}'", textBox1.Text);
-                dataGridView1.DataSource = getData(sql2);
+                string sql = string.Format("delete  from 公寓寝室 where 公寓号='{0}' and 寝室号='{1}'", dataGridView1.SelectedCells[0].Value.ToString(), dataGridView1.SelectedCells[1].Value.ToString());
+                int i = Sqlhelper.exexute(sql);
+                if (i == 1)
+                {
+                    MessageBox.Show("删除成功！");
+                    string sql2 = string.Format("select * from 公寓寝室 where 寝室号='{0}'", textBox1.Text);
+                    dataGridView1.DataSource = getData(sql2);
+                }
+                else
+                {
+                    MessageBox.Show("删除失败！");
+                }
             }
-            else
-            {
-                MessageBox.Show("删除失败！");
-            }
+            
         }
     }
 }
